Split phase space plotting into new method
Previous code had grown over time so that there was lots of duplication for the three sections for initial, final and BPM plots. This change moves that code into a new common method, `plot_phase_spaces_and_energies()`.

It also adds some extra error handling to make sure the maximum amount of plots can be generated when run as a post-processing script.