Get `z_offset` from input file rather than requiring user input
When run as a script, the offset is loaded from file at the beginning of the `plot_all()` routine.

When run as a GUI, the z-offset is loaded at draw time as part of the `create_offset_selector()` method, and the `get_offset()` method is also updated to check the input file if there is no value set in the GUI itself.